backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
motor==3.3.2
pymongo==4.6.0
pydantic==2.5.2
openai==1.3.7
python-dotenv==1.0.0
